Fix rule dedup() returning None
userIdentity.arn may be missing in some CloudTrail events.
Returning None by dedup() raises exception (we assert title/dedup
return str), which is logged but the system behavior is the same
(default value is used for title/dedup).

The result is spamming the logs (which may trigger CloudWatch alarms)
with no benefit.