New Notion Detections (#896)
* new rule: login from new location

* new rule: password changed after login

* made pylint happy(er)

* new rule: password removed shortly after login

* new rule: email changed shortly after login

* Fix: notion wraps all fields in "event" object

* Impossible travel, and minor changes

* Fixed small type in data model

* More fixes to data model

* Impossible travel for page views

* Added detection for pages made public

* Added alert_context to Notion.PageSharedToWeb

* new rule: integration added to page

* edit rule: forgot to change RuleID

* new rules: guests added to page

* Fixed Notion data model for event type

* Added Notion to std travel rule, fixed data model

* Combined email+password change rules into 1 rule

* remove rule: team guest added to page

* new rule: login from blocked IP

* fix: minor linting and bug fixes, pass 1

* fix: linting and bug stuff pass 2

* fix: better title for many pages exported

* fix: improper deep get in title for scim rule

* fix: minor tweaks to title() and alert_context()

* fix: made linter happy again

* edit: notion_login_from_new_location

Combine TTL assignment and put_string_set to one call

Co-authored-by: darwayne <darwaynelynch@gmail.com>

* edit: notion_login_from_new_location

Force TTL check when checking for past logins

Co-authored-by: darwayne <darwaynelynch@gmail.com>

* edit: account changed after login

Changed the default assignment of LOGIN_TS to the module scope.

* fix: linting

* fix: updated pack to include new detections

* edit: changed caching method

Using dictionaries instead of string sets means we can detect if this is
a user's first login, and if so, then not alert. Makes the rule less noisy
when turned on for the first time.

* edit: changed phrasing and added tests to 2 rules

Based on feedback from Notion about what the page permission
events mean, we adjusted the names/IDs of the related rules. Also
added some unit tests and tweaked the titles/alert_contexts.

* fix: title wording for SAML SSO config change rule

The SSO change is applied globally, not to a single workspace. The
workspace ID in the event payload informs the user which workspace
the action was performed from, not which workspace it applies to.
(Based on a conversation with Notion.)

* edit: standardized tags used across notion rules

* format: sorted keys in the YAML files

For consistency, I ordered the YAML keys according to the format we
output when exporting rules from the UI.

* fix: no enrichment in login from new location

Notion.LoginFromNewLocation would cause an error if there wasn't
any location enrichment - now it will return False if that's the case.

* edit: remove some rules from pack

Removed the impossible travel page view rule, and the login
from blocked IP rule. Both of these are noisy unless tuned, so we
want to avoind them being enabled when turning on the whole pack.

---------

Co-authored-by: darwayne <darwaynelynch@gmail.com>