Update unittest call syntax
When a glob is passed directly to `python`, and there are two or more
matched files, the 2nd+ is considered an argument to the program.
Unittest then interprets this argument as a function name to look for,
and fails with an error like the following:

  ======================================================================
  ERROR: global_helpers (unittest.loader._FailedTest)
  ----------------------------------------------------------------------
  AttributeError: module '__main__' has no attribute 'global_helpers'

  ----------------------------------------------------------------------
  Ran 1 test in 0.000s

However, if multiple filenames from a glob are passed to
`python -m unittest`, then all the intended test cases are run.