Create Github.Repo.RulesetModified rule
### Background

The cool kidz these days are using ruleset instead of or in combination with branch protection rules. This rule alerts to modifications (update, destroy) of rulesets and adds an INFO-level signal for ruleset creations if desired for future correlation rules. Filters can be applied in global_filter_github and github_alert_context is enriched specific to repo ruleset changes.

[1] https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-rulesets/about-rulesets

### Changes

- Create GitHub.Repo.RulsetModified rule
- Tests for each create/update/destroy event type
- Update the github.yml pack

### Testing

pat test, make lint